% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplot.R
\name{categorical-color}
\alias{categorical-color}
\alias{fcolor}
\alias{hcld.colors}
\alias{cat.colors}
\title{Utilities for plotting with a categorical legend}
\usage{
fcolor(f, col = hcld.colors(length(labels)), labels = levels(as.factor(f)))

hcld.colors(n, palette = "Dark 3", ...)

cat.colors(n)
}
\arguments{
\item{f}{(factor, or vector coercible to factor) values to be converted to colors.}

\item{col}{colors for each level. Defaults to \code{hcld.colors()}.}

\item{labels}{levels to use. Defaults to \code{levels(as.factor(f))}.}

\item{n}{number of colors (\code{>= 1}) to be in the palette.}

\item{palette}{a valid palette name for \code{hcl.colors()} (one of \code{hcl.pals()}).}

\item{...}{additional arguments passed to \code{legend()}.}
}
\value{
\code{fcolor()} returns vector of colors, one per element of \code{f}.
\code{hcld.colors()} and \code{cat.colors()} return a character vector of colors (similar to
\code{hcl.colors()} or \code{rainbow()}; see \code{rgb()}).
}
\description{
\code{hcld.colors()} and \code{hot.colors()} create a color table useful for
coding qualitative information and \code{fcolor()} assigns colors to categorical
(factor) values.
}
