% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splot.R
\name{continuous-color}
\alias{continuous-color}
\alias{scolor}
\alias{jet.colors}
\alias{hot.colors}
\title{Utilities for plotting with a continuous color scale}
\usage{
scolor(s, col = jet.colors(128), slim = range(s, finite = TRUE))

jet.colors(n)

hot.colors(n, rev = TRUE)
}
\arguments{
\item{s}{values to be converted to the color scale.}

\item{col}{color table used to set up the color scale (see \code{\link{image}} for
details).}

\item{slim}{limits used to set up the color scale.}

\item{n}{number of colors (\code{>= 1}) to be in the palette.}

\item{rev}{logical; if \code{TRUE}, the palette is reversed (decreasing overall luminosity).}
}
\value{
\code{scolor}, \code{jet.colors} and \code{hot.colors} return a character
vector of colors.
}
\description{
\code{jet.colors} and \code{hot.colors} create a color table useful for contiguous
color scales and \code{scolor} assigns colors to a numerical vector.
}
\details{
\code{scolor} converts a real valued vector to a color scale. The range
\code{slim} is divided into \code{length(col) + 1} pieces of equal length.
Values which fall outside the range of the scale are coded as \code{NA}.

\code{jet.colors} generates a rainbow style color table similar to the MATLAB (TM)
jet color scheme. It may be appropriate to distinguish between values above and
below a central value (e.g. between positive and negative values).

\code{hot.colors} generates a color table similar to the MATLAB (TM)
hot color scheme (reversed by default). It may be appropriate to represent values
ranging from 0 to some maximum level (e.g. density estimation).
The default value \code{rev = TRUE} may be adequate to grayscale conversion.
}
\seealso{
\code{\link{heat.colors}}, \code{\link{terrain.colors}}, \code{\link{rgb}}.
}
