% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpoints.R
\name{fpoints}
\alias{fpoints}
\alias{fpoints.default}
\title{Scatter plot with a categorical legend}
\usage{
fpoints(x, ...)

\method{fpoints}{default}(
  x,
  y = NULL,
  f,
  col = hcld.colors(f),
  type = "p",
  cex = 1.5,
  pch = 16,
  legend = TRUE,
  legend.type = c("point", "box", "line"),
  legend.pch = pch,
  legend.cex = 1,
  legend.lty = 1,
  legend.lwd = 2,
  border = col,
  pt.cex = cex,
  seg.len = 1.5,
  horizontal = FALSE,
  legend.shrink = 1,
  legend.width = NULL,
  legend.mar = NULL,
  legend.lab = NULL,
  legend.x = "center",
  bigplot = NULL,
  smallplot = NULL,
  add = FALSE,
  reset = TRUE,
  xlab = NULL,
  ylab = NULL,
  asp = NA,
  ...
)
}
\arguments{
\item{x}{object used to select a method. In the default method, it provides the \code{x}
coordinates for the plot (and optionally the \code{y} coordinates;
any reasonable way of defining the coordinates is acceptable,
see the function \code{\link[=xy.coords]{xy.coords()}} for details).}

\item{...}{additional graphical parameters (to be passed to the main plot function
or to \code{fpoints.default()}; e.g. \verb{xlim, ylim,} ...). NOTE:
graphical arguments passed here will only have impact on the main plot.
To change the graphical defaults for the legend use the \code{\link[=par]{par()}}
function beforehand.}

\item{y}{y coordinates. Alternatively, a single argument \code{x} can be provided.}

\item{f}{(factor, or vector coercible to factor), with length equal to the
number of points, giving the group of each point.}

\item{col}{colors associated with each level of \code{f} Defaults to \code{hcld.colors()}.}

\item{type}{character indicating the type of plotting; actually any of the
types as in \code{\link[=plot.default]{plot.default()}}.}

\item{cex}{numerical vector giving the amount by which plotting characters
and symbols should be scaled relative to the default. This works as a multiple
of \code{par("cex")}.}

\item{pch}{vector of plotting characters or symbols: see \code{\link[=points]{points()}}.}

\item{legend}{logical; if \code{TRUE} (default), the plotting region is
splitted into two parts, drawing the scatter plot in one and the
categorical legend in the other. If \code{FALSE} only the (coloured)
scatter plot is drawn and the legend-related arguments are ignored
(\code{\link[=fplot]{fplot()}} is not called).}

\item{legend.type}{ype of symbols shown in the legend: \code{"box"} for filled
color boxes (as in a classic factor-level legend), \code{"point"} for points, or
\code{"line"} for line segments. (see \code{\link[=fplot]{fplot()}}).}

\item{legend.pch, legend.cex}{plotting character and size used in the legend
when \code{legend.type = "point"}.}

\item{legend.lty, legend.lwd}{line type and width used in the legend
when \code{legend.type = "line"}.}

\item{border, pt.cex, seg.len}{additional \code{\link[=legend]{legend()}} parameters,
also used to estimate the required legend width/height.
The default values are: \code{border = col}, \code{pt.cex = cex * 1.5} and \code{seg.len = 1.5}.}

\item{horizontal}{logical; if \code{FALSE} (default) legend will appear on the right side.
If \code{TRUE} the legend will be along the bottom.}

\item{legend.shrink}{amount to shrink the size of legend relative to the
full height or width of the plot.}

\item{legend.width}{width in characters of the legend strip. Default is
1.2, a little bigger that the width of a character.}

\item{legend.mar}{width in characters of legend margin that has the axis.
Default is 5.1 for a vertical legend and 3.1 for a horizontal legend.}

\item{legend.lab}{label for the axis of the color legend, defaults to a
description of \code{f}.}

\item{legend.x}{legend location relative to the legend panel (argument \code{x}
of \code{\link[=legend]{legend()}}). Possible values are: \code{"center"} (default), \code{"bottomright"},
\code{"bottom"}, \code{"bottomleft"}, \code{"left"}, \code{"topleft"}, \code{"top"}, \code{"topright"} or \code{"right"}.}

\item{bigplot}{plot coordinates for main plot. If not passed, and \code{legend}
is TRUE, these will be determined within the function.}

\item{smallplot}{plot coordinates for legend strip. If not passed, and \code{legend}
is TRUE, these will be determined within the function.}

\item{add}{logical; if \code{TRUE} the scatter plot is just added
to the existing plot (including the legend if \code{legend = TRUE}, although the
graphical parameters are not modified).}

\item{reset}{logical; if \code{FALSE} the plotting region
(\code{par("plt")}) will not be reset to make it possible to add more features
to the plot (e.g. using functions such as points or lines). If \code{TRUE} (default)
the plot parameters will be reset to the values before entering the function.}

\item{xlab}{label for the x axis, defaults to a description of \code{x}.}

\item{ylab}{label for the y axis, defaults to a description of \code{y}.}

\item{asp}{the y/x aspect ratio, see \code{\link[=plot.window]{plot.window()}}.}
}
\value{
\code{fplot()} invisibly returns a list with components: \code{bigplot}, \code{smallplot},
\code{old.par}, \code{col} and \code{labels} (\code{par(old.par)} will reset plot
parameters to the values before entering the function).
}
\description{
A generic function that, by default, draws a scatter plot with
points colored according to a factor \code{f} and  (optionally) adds a categorical
legend (\code{fpoints.default()} calls \code{\link[=fplot]{fplot()}} and \code{\link[=plot.default]{plot.default()}}, or
\code{\link[=plot.xy]{plot.xy()}} if \code{add = TRUE}).
}
\section{Side Effects}{

If \code{reset = TRUE}, the plotting region (\code{par("plt")})
may be changed after exiting, to make it possible to add more features to the
plot (setting \code{reset = FALSE} prevents this).
}

\examples{
with(mtcars,
    fpoints(hp, qsec, f = cyl, main = "Motor Trend Car Road Tests")
)

}
\seealso{
\code{\link[=fplot]{fplot()}}, \code{\link[=fcolor]{fcolor()}}, \code{\link[=hcld.colors]{hcld.colors()}}, \code{\link[=plot.default]{plot.default()}}, \code{\link[=spoints]{spoints()}}.
}
\keyword{hplot}
