% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpoints3d.R
\name{fpoints3d}
\alias{fpoints3d}
\alias{fpoints3d.default}
\title{3D scatter plot with a categorical legend}
\usage{
fpoints3d(x, ...)

\method{fpoints3d}{default}(
  x,
  y = NULL,
  z = NULL,
  f,
  col = hcld.colors(nlevels(f)),
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  type = "p",
  legend = TRUE,
  legend.type = c("s", "c", "p", "l"),
  legend.zoom = 0.6,
  legend.width = 0.2,
  legend.size = 4,
  legend.dim = 0.2,
  legend.lab = NULL,
  lab.dist = 2.5,
  lab.rev = FALSE,
  add = FALSE,
  ...
)
}
\arguments{
\item{x}{object used to select a method. In the default method, it provides the \code{x}
coordinates for the plot (and optionally the \code{y} and \code{z} coordinates;
any reasonable way of defining the coordinates is acceptable,
see the function \code{\link[=xyz.coords]{xyz.coords()}} for details).}

\item{...}{additional arguments passed to \code{\link[rgl:plot3d]{rgl::plot3d()}} for the main
scatter plot (e.g. \code{size}).}

\item{y, z}{\code{y} and \code{z} point coordinates. Alternatively, a single argument \code{x}
can be provided.}

\item{f}{(factor, or vector coercible to factor), used to color the points.}

\item{col}{vector of colors associated with each level of \code{f} (defaults to
\code{hcld.colors(nlevels(f))}).}

\item{xlab, ylab, zlab}{labels for the coordinates.}

\item{type}{character indicating the type of item to plot (see
\code{\link[rgl:plot3d]{rgl::plot3d()}}: \code{"p"} for points, \code{"s"} for spheres, \code{"l"} for lines,
\code{"h"} for line segments from z = 0, and \code{"n"} for nothing).}

\item{legend}{logical; if \code{TRUE} (default), the active rgl device is splitted
into two subscenes, drawing the main plot in one and the categorical
legend in the other (see \code{\link[=fplot3d]{fplot3d()}}).
if \code{FALSE} only the (coloured) main plot is drawn and the arguments related to
the legend are ignored (\code{\link[=fplot3d]{fplot3d()}} is not called).}

\item{legend.type}{character; symbol/object used to represent each
level in the legend (see \code{type} in \code{\link[=fplot3d]{fplot3d()}}): \code{"s"} for spheres
(default), \code{"c"} for cubes, \code{"p"} for points, or \code{"l"} for
(horizontal) line segments.}

\item{legend.zoom}{zoom factor applied to the legend subscene/panel.}

\item{legend.width}{relative size of the legend symbols (spheres, cubes or
line segments), as a fraction of the corresponding dimension
of the legend subscene.}

\item{legend.size}{point size, if \code{type = "p"}, or line width, if \code{type = "l"}
(in pixels; see \code{size} and \code{lwd} graphical parameters in \code{\link[rgl:par3d]{rgl::par3d()}};
ignored in all other cases).}

\item{legend.dim}{relative dimension of the legend panel, as a fraction of
the full width (or height) of the device.}

\item{legend.lab}{label for the axis of the color legend, defaults to a
description of \code{f}.}

\item{lab.dist}{distance between the legend symbols and their text
labels, as a multiple of the symbol size (\code{legend.width}).}

\item{lab.rev}{logical; if \code{TRUE} the order of the levels in the legend
is reversed (by default they are shown from top to bottom).}

\item{add}{logical; if \code{TRUE} the scatter plot is just added to the existing
plot (including the legend if \code{legend = TRUE}).}
}
\value{
Called for its side effect (draws the 3D scatter plot and, unless
\code{legend = FALSE}, the legend on the active rgl device).
}
\description{
A generic function that, by default, draws a 3D scatter plot (\code{\link[rgl:plot3d]{rgl::plot3d()}})
with points colored according to the levels of a factor \code{f}, and (optionally)
adds a categorical legend (via \code{\link[=fplot3d]{fplot3d()}}).
}
\examples{
library(rgl)
open3d() # Alternatively, use `new3d()` to clear the current device or open a new one
with(mtcars, fpoints3d(hp, qsec, mpg, f = cyl, type = "s"))

}
\seealso{
\code{\link[=spoints3d]{spoints3d()}}, \code{\link[=fplot3d]{fplot3d()}}, \code{\link[=fcolor]{fcolor()}}, \code{\link[=fshade3d]{fshade3d()}}, \code{\link[rgl:plot3d]{rgl::plot3d()}}.
}
