% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendplot_package.R
\docType{package}
\name{legendplot-package}
\alias{legendplot-package}
\alias{legendplot}
\title{legendplot: Standard and 'rgl' Plots with Legends}
\description{
This package provides tools to combine standard R plots or 'rgl' 3D plots
with a legend, facilitating the creation of composite figures that mix 2D or 3D
visualizations with a categorical or continuous legend.
For more information visit \url{https://rubenfcasal.github.io/legendplot/articles/legendplot.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rubenfcasal.github.io/legendplot/}
  \item \url{https://github.com/rubenfcasal/legendplot/}
  \item Report bugs at \url{https://github.com/rubenfcasal/legendplot/issues/}
}

}
\author{
\strong{Maintainer}: Ruben Fernandez-Casal \email{rubenfcasal@gmail.com} (\href{https://orcid.org/0000-0002-5785-3739}{ORCID})

}
