% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.reset.R
\name{par.reset}
\alias{par.reset}
\title{Restore graphical parameters}
\usage{
par.reset(default = FALSE, set = FALSE)
}
\arguments{
\item{default}{logical; if \code{FALSE} (default) restores the parameters saved
automatically by the last call to \code{\link[=splot]{splot()}} or \code{\link[=fplot]{fplot()}}. If \code{TRUE},
restores the \emph{default} parameters (those in effect before the first call
to \code{\link[=splot]{splot()}}/\code{\link[=fplot]{fplot()}} or set with \code{par.reset(set = TRUE)}).}

\item{set}{logical; if \code{TRUE}, saves the \emph{current} graphical parameters
as the new defaults (the graphical parameters are not changed).}
}
\value{
Invisibly returns the graphical parameters as they were before
restoring/saving (see \code{\link[=par]{par()}}).
}
\description{
Restores the graphical parameters changed by \code{\link[=splot]{splot()}}/\code{\link[=fplot]{fplot()}} (or by
\code{sxxx()}/\code{fxxx()} functions with \code{reset = FALSE}) or those previously set
as the default ones.
}
\examples{
scale.range <- range(mtcars$mpg)
splot(slim = scale.range, legend.lab = "mpg")
with(mtcars,
  plot(hp, qsec, col = scolor(mpg, slim = scale.range),
       pch = 16, cex = 1.5)
)
par.reset() # restores parameters from the last splot()/fplot() call

# Set current parameters as the new default
par.reset(set = TRUE)
# ... later, after several splot()/fplot() calls ...
par.reset(default = TRUE) # restores those that were set as defaults

}
\seealso{
\code{\link[=splot]{splot()}}, \code{\link[=fplot]{fplot()}}, \code{\link[=par]{par()}}
}
