% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spersp3d.R
\name{spersp3d}
\alias{spersp3d}
\alias{spersp3d.default}
\title{3D surface plot with a continuous color scale}
\usage{
spersp3d(x, ...)

\method{spersp3d}{default}(
  x,
  y = NULL,
  z = NULL,
  s = z,
  slim = range(s, finite = TRUE),
  col = jet.colors(128),
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  xlim = NULL,
  ylim = NULL,
  zlim = NULL,
  aspect = !add,
  legend = TRUE,
  legend.zoom = 0.4,
  legend.width = 0.1,
  legend.dim = 0.2,
  legend.lab = NULL,
  box = TRUE,
  lab.breaks = NULL,
  lab.ticksize = 0.5,
  lab.dist = 3,
  add = FALSE,
  ...
)
}
\arguments{
\item{x}{object used to select a method. In the default method, it typically
provides the grid values for the \code{x} axis. The grid may be specified in
several ways, see \code{\link[rgl:surface3d]{rgl::surface3d()}}.}

\item{...}{additional arguments passed to \code{\link[rgl:persp3d]{rgl::persp3d()}} for the main plot.}

\item{y}{grid values for the \code{y} axis.}

\item{z}{matrix with the surface height at each grid point (of dimension
\code{length(x)} by \code{length(y)} if \code{x} and \code{y} are vectors).}

\item{s}{matrix with the values used for coloring the surface (one color per
vertex). Defaults to \code{z}.}

\item{slim}{limits (vector of length 2 with the minimum and maximum) used
to set up the color scale.}

\item{col}{color table used to set up the color scale. Defaults to
\code{jet.colors(128)}.}

\item{xlab, ylab, zlab}{titles for the axes (character strings; expressions are
not accepted)}

\item{xlim, ylim, zlim}{x-, y- and z-limits.  If present, the plot is clipped
to this region.}

\item{aspect}{either a logical indicating whether to adjust the aspect
ratio, or a new ratio.}

\item{legend}{logical; if \code{TRUE} (default), the active rgl device is splitted
into two subscenes, drawing the main plot in one and the categorical
legend in the other (see \code{\link[=fplot3d]{fplot3d()}}).
if \code{FALSE} only the (coloured) main plot is drawn and the arguments related to
the legend are ignored (\code{\link[=fplot3d]{fplot3d()}} is not called).}

\item{legend.zoom}{zoom factor applied to the legend subscene (see
\code{\link[rgl:view3d]{rgl::view3d()}}).}

\item{legend.width}{relative width of the color bar, as a fraction of
the corresponding dimension of the legend subscene/panel.}

\item{legend.dim}{relative dimension of the legend panel, as a fraction of
the full width (or height) of the device.}

\item{legend.lab}{legend title.}

\item{box}{logical; if \code{TRUE} (the default) a rectangle is drawn around
the color bar.}

\item{lab.breaks}{labels for the legend axis breaks; if \code{NULL} they are
generated automatically (see \code{\link[=axis3]{axis3()}}).}

\item{lab.ticksize}{length of the legend axis tick marks, as a fraction
of the distance to the center of the legend panel.}

\item{lab.dist}{distance between the tick marks and the labels of the
legend axis, as a multiple of \code{lab.ticksize}.}

\item{add}{logical; if \code{TRUE} the scatter plot is just added to the existing
plot (including the legend if \code{legend = TRUE}).}
}
\value{
Called for its side effect (draws the 3D surface plot and, unless
\code{legend = FALSE}, the legend on the active rgl device).
}
\description{
A generic function that draws a 3D surface (\code{\link[rgl:persp3d]{rgl::persp3d()}})
colored according to a continuous scale associated with a vector
of values \code{s}, and (optionally) adds a color-bar legend (via \code{\link[=splot3d]{splot3d()}}).
}
\examples{
library(rgl)
open3d() # Alternatively, use `new3d()` to clear the current device or open a new one
x <- seq(0, 1, length.out = 30)
y <- seq(0, 1, length.out = 30)
z <- outer(x, y, function(x, y) sin(2*pi*x) + 4*(y-0.5)^2 - 0.5)
spersp3d(x, y, z)

}
\seealso{
\code{\link[=splot3d]{splot3d()}}, \code{\link[=scolor]{scolor()}}, \code{\link[=sshade3d]{sshade3d()}}, \code{\link[rgl:persp3d]{rgl::persp3d()}}.
}
