% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendplot_package.R
\docType{data}
\name{volcanom}
\alias{volcanom}
\title{Surface Mesh of Auckland's Maunga Whau Volcano}
\format{
\subsection{\code{volcanom}}{

A \link[rgl:mesh3d]{rgl::mesh3d}-class object with 5307 vertices and 10320 triangles.
}
}
\source{
\link[datasets:volcano]{datasets::volcano} and example taken from \link[rgl:surface3d]{rgl::surface3d}.
}
\usage{
volcanom
}
\description{
Surface mesh corresponding to the altitude on a 10m by 10m grid
of Auckland's Maunga Whau volcano.
}
\examples{
library(rgl)
open3d()
aspect3d(1, 1, 3)  # Exaggerate the relief
shade3d(volcanom, col = "lightgreen")
}
\keyword{datasets}
