% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new3d.R
\name{new3d}
\alias{new3d}
\title{New 'rgl' plot}
\usage{
new3d(open = FALSE, clear = "all", ...)
}
\arguments{
\item{open}{if \code{TRUE} a new rgl device is opened.}

\item{clear}{rgl stack(s) to remove, see argument \code{type} in \code{\link[rgl:scene]{rgl::clear3d()}}
(used only if there is an active device and \code{open = FALSE}).
Defaults to \code{"all"}, with a new light source added to the scene.}

\item{...}{additional arguments, passed to \code{\link[rgl:open3d]{rgl::open3d()}} or \code{\link[rgl:scene]{rgl::clear3d()}}.}
}
\value{
Called for its side effect (opens or clears the rgl device, and configures
the mouse mode); invisibly returns the current device.
}
\description{
Clear the current rgl device or open a new one (if there are no devices or
if \code{open = TRUE}).
Additionally assign the middle button to zoom (via \code{\link[=setmousemode]{setmousemode()}}) and
the right button to panning (via \code{\link[=pan3d]{pan3d()}}).
}
\examples{
library(rgl)
# New rgl device
new3d()
shade3d(volcanom, col = "lightgreen")
# Use the right mouse button to panning and the middle button to zoom
# ...
# Clear the current device
new3d()
sshade3d(volcanom, s = volcanom$vb[3, ], meshColor = "facesvertices",
         col = terrain.colors(128))
}
\seealso{
\code{\link[rgl:open3d]{rgl::open3d()}}, \code{\link[rgl:scene]{rgl::clear3d()}}
}
