% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simage.R
\name{simage}
\alias{simage}
\alias{simage.default}
\title{Image plot with a color scale}
\usage{
simage(x, ...)

\method{simage}{default}(
  x = seq(0, 1, len = nrow(s)),
  y = seq(0, 1, len = ncol(s)),
  s,
  slim = range(s, finite = TRUE),
  col = jet.colors(128),
  breaks = NULL,
  legend = TRUE,
  horizontal = FALSE,
  legend.shrink = 0.8,
  legend.width = 1.2,
  legend.mar = ifelse(horizontal, 3.1, 5.1),
  legend.lab = NULL,
  bigplot = NULL,
  smallplot = NULL,
  lab.breaks = NULL,
  axis.args = NULL,
  legend.args = NULL,
  reset = TRUE,
  xlab = NULL,
  ylab = NULL,
  asp = NA,
  ...
)
}
\arguments{
\item{x}{grid values for \code{x} coordinate. If \code{x} is a list,
its components \code{x$x} and \code{x$y} are used for \code{x}
and \code{y}, respectively. For compatibility with \code{\link{image}}, if the
list has component \code{z} this is used for \code{s}.}

\item{...}{additional graphical parameters (to be passed to \code{\link{image}}
or \code{simage.default}; e.g. \code{xlim, ylim,} ...). NOTE:
graphical arguments passed here will only have impact on the main plot.
To change the graphical defaults for the legend use the \code{\link{par}}
function beforehand (e.g. \code{par(cex.lab = 2)} to increase colorbar labels).}

\item{y}{grid values for \code{y} coordinate.}

\item{s}{matrix containing the values to be used for coloring the rectangles (NAs are allowed).
Note that \code{x} can be used instead of \code{s} for convenience.}

\item{slim}{limits used to set up the color scale.}

\item{col}{color table used to set up the color scale (see \code{\link{image}} for
details).}

\item{breaks}{(optional) numeric vector with the breakpoints for the color scale:
must have one more breakpoint than \code{col} and be in increasing order.}

\item{legend}{logical; if \code{TRUE} (default), the plotting region is splitted into two parts,
drawing the image plot in one and the legend with the color scale in the other.
If \code{FALSE} only the image plot is drawn and the arguments related
to the legend are ignored (\code{\link{splot}} is not called).}

\item{horizontal}{logical; if \code{FALSE} (default) legend will appear on the right side.
If \code{TRUE} the legend will be along the bottom.}

\item{legend.shrink}{amount to shrink the size of legend relative to the
full height or width of the plot.}

\item{legend.width}{width in characters of the legend strip. Default is
1.2, a little bigger that the width of a character.}

\item{legend.mar}{width in characters of legend margin that has the axis.
Default is 5.1 for a vertical legend and 3.1 for a horizontal legend.}

\item{legend.lab}{label for the axis of the color legend. Default is no
label as this is usual evident from the plot title.}

\item{bigplot}{plot coordinates for main plot. If not passed these will be
determined within the function.}

\item{smallplot}{plot coordinates for legend strip. If not passed these
will be determined within the function.}

\item{lab.breaks}{if breaks are supplied these are text string labels to
put at each break value. This is intended to label axis on a transformed
scale such as logs.}

\item{axis.args}{additional arguments for the axis function used to create
the legend axis (see \code{\link[fields]{image.plot}} for details).}

\item{legend.args}{arguments for a complete specification of the legend
label. This is in the form of list and is just passed to the \code{\link{mtext}}
function. Usually this will not be needed (see \code{\link[fields]{image.plot}}
for details).}

\item{reset}{logical; if \code{FALSE} the plotting region
(\code{par("plt")}) will not be reset to make it possible to add more features
to the plot (e.g. using functions such as points or lines). If \code{TRUE} (default)
the plot parameters will be reset to the values before entering the function.}

\item{xlab}{label for the x axis, defaults to a description of \code{x}.}

\item{ylab}{label for the y axis, defaults to a description of \code{y}.}

\item{asp}{the y/x aspect ratio, see \code{\link[=plot.window]{plot.window()}}.}
}
\value{
Invisibly returns a list with the following 3 components:
\item{bigplot}{plot coordinates of the main plot. These values may be useful for
drawing a plot without the legend that is the same size as the plots with legends.}
\item{smallplot}{plot coordinates of the secondary plot (legend strip).}
\item{old.par}{previous graphical parameters (\code{par(old.par)}
will reset plot parameters to the values before entering the function).}
}
\description{
\code{simage} (generic function) draws an image (a grid of colored rectangles)
and (optionally) adds a legend strip with the color scale
(calls \code{\link{splot}} and \code{\link{image}}).
}
\section{Side Effects}{

If \code{reset = FALSE}, the plotting region (\code{par("plt")}) may be changed
after exiting, to make it possible to add more features to the plot.
They can be restored using the \code{old.par} returned values or by calling
function \code{\link[=par.reset]{par.reset()}}.
}

\examples{
# Regularly spaced 2D data
nx <- c(40, 40) # ndata =  prod(nx)
x1 <- seq(-1, 1, length.out = nx[1])
x2 <- seq(-1, 1, length.out = nx[2])
trend <- outer(x1, x2, function(x,y) x^2 - y^2)
simage( x1, x2, trend, main = 'Trend')
}
\seealso{
\code{\link{splot}}, \code{\link{spoints}}, \code{\link{spersp}},
\code{\link{image}}, \code{\link[fields]{image.plot}}.
}
\author{
Based on \code{\link[fields]{image.plot}} function from package \pkg{fields}:
fields, Tools for spatial data.
Copyright 2004-2013, Institute for Mathematics Applied Geosciences.
University Corporation for Atmospheric Research.

Modified by Ruben Fernandez-Casal \href{mailto:rubenfcasal@gmail.com}{rubenfcasal@gmail.com}.
}
\keyword{hplot}
